requirements.txt
python-telegram-bot==13.15
pyyaml==6.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3 